DOSE-package {DOSE}	R Documentation

Disease Ontology Semantic and Enrichment analysis Implemented five methods proposed by Resnik, Schlicker, Jiang, Lin and Wang respectively for measuring DO semantic similarities, and hypergeometric test for enrichment analysis.

Description

This package is designed to estimate DO-based semantic similarity measurement and enrichment analysis.

Details

Package: 	 DOSE

Type: 	 Package

Version: 	 2.3.5

Date: 	 2-27-2012

biocViews:	 Bioinformatics, Annotation

Depends:

Imports: 	 methods, AnnotationDbi, DO.db

Suggests:	 clusterProfiler, GOSemSim

License: 	 Artistic-2.0

Author(s)

Guangchuang Yu, Li-Gen Wang

Maintainer: Guangchuang Yu guangchuangyu@gmail.com

See Also

enrichResult

[Package DOSE version 3.20.1 Index]