GSEA_internal {DOSE}	R Documentation

GSEA_internal

Description

generic function for gene set enrichment analysis

Usage

GSEA_internal(
  geneList,
  exponent,
  minGSSize,
  maxGSSize,
  eps,
  pvalueCutoff,
  pAdjustMethod,
  verbose,
  seed = FALSE,
  USER_DATA,
  by = "fgsea",
  ...
)

Arguments

geneList
order ranked geneList

exponent
weight of each step

minGSSize
minimal size of each geneSet for analyzing

maxGSSize
maximal size of each geneSet for analyzing

eps
This parameter sets the boundary for calculating the p value.

pvalueCutoff
p value Cutoff

pAdjustMethod
p value adjustment method

verbose
print message or not

seed
set seed inside the function to make result reproducible. FALSE by default.

USER_DATA
annotation data

by
one of 'fgsea' or 'DOSE'

...
other parameter

Value

gseaResult object

Author(s)

Yu Guangchuang

[Package DOSE version 3.20.1 Index]