compareClusterResult-class {DOSE}	R Documentation

Class "compareClusterResult" This class represents the comparison result of gene clusters by GO categories at specific level or GO enrichment analysis.

Description

Class "compareClusterResult" This class represents the comparison result of gene clusters by GO categories at specific level or GO enrichment analysis.

Slots

compareClusterResult

cluster comparing result

geneClusters

a list of genes

fun

one of groupGO, enrichGO and enrichKEGG

gene2Symbol

gene ID to Symbol

keytype

Gene ID type

readable

logical flag of gene ID in symbol or not.

.call

function call

termsim

Similarity between term

method

method of calculating the similarity between nodes

dr

dimension reduction result

Author(s)

Guangchuang Yu https://guangchuangyu.github.io

See Also

enrichResult

[Package DOSE version 3.20.1 Index]