enrichNCG {DOSE}	R Documentation

enrichNCG

Description

Enrichment analysis based on the Network of Cancer Genes database (http://ncg.kcl.ac.uk/)

Usage

enrichNCG(
  gene,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  readable = FALSE
)

Arguments

gene
a vector of entrez gene id

pvalueCutoff
pvalue cutoff

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

universe
background genes

minGSSize
minimal size of genes annotated by NCG category for testing

maxGSSize
maximal size of each geneSet for analyzing

qvalueCutoff
qvalue cutoff

readable
whether mapping gene ID to gene Name

Details

given a vector of genes, this function will return the enrichment NCG categories with FDR control

Value

A enrichResult instance

Author(s)

Guangchuang Yu

[Package DOSE version 3.20.1 Index]