geneSim {DOSE}	R Documentation

geneSim

Description

measuring similarities bewteen two gene vectors.

Usage

geneSim(geneID1, geneID2 = NULL, measure = "Wang", combine = "BMA")

Arguments

geneID1
entrez gene vector

geneID2
entrez gene vector

measure
one of "Wang", "Resnik", "Rel", "Jiang", and "Lin".

combine
One of "max", "avg", "rcmax", "BMA" methods, for combining semantic similarity scores of multiple DO terms associated with gene/protein.

Details

provide two entrez gene vectors, this function will calculate their similarity.

Value

score matrix

Author(s)

Guangchuang Yu http://ygc.name

[Package DOSE version 3.20.1 Index]