mclusterSim {DOSE}	R Documentation

mclusterSim

Description

Pairwise semantic similarity for a list of gene clusters

Usage

mclusterSim(clusters, measure = "Wang", combine = "BMA")

Arguments

clusters
A list of gene clusters

measure
one of "Wang", "Resnik", "Rel", "Jiang", and "Lin".

combine
One of "max", "avg", "rcmax", "BMA" methods, for combining semantic similarity scores of multiple DO terms associated with gene/protein.

Value

similarity matrix

Author(s)

Yu Guangchuang

Examples

cluster1 <- c("835", "5261","241")
cluster2 <- c("578","582")
cluster3 <- c("307", "308", "317")
clusters <- list(a=cluster1, b=cluster2, c=cluster3)
mclusterSim(clusters, measure="Wang")

[Package DOSE version 3.20.1 Index]